requirements.txt
aiogram>=3.27.0,<4.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
jinja2>=3.1.4,<4.0.0
pydantic>=2.8.0,<3.0.0

